divideModal
